% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.plot.R
\name{data_cleaning}
\alias{data_cleaning}
\title{This function cleans all of the NA values from a data set, as well as creates a scatter plot using designated x and y values and saves the plot to a file.}
\usage{
data_cleaning(filepath)
}
\arguments{
\item{filepath}{the name of a dataset that has been loaded into the
environment or its corresponding file path}

\item{savefile}{the name that your plot file will be saved as}

\item{x}{= x value}

\item{y}{= y value}
}
\value{
If data cleaning is unsuccessful, console will print "NAs still present!".

If non-integer values are entered for the plot function, console will print "Error: Values must be integers!"
}
\description{
This function cleans all of the NA values from a data set, as well as creates a scatter plot using designated x and y values and saves the plot to a file.
}
\examples{
data_cleaning(dataset_name)
create_plot(dataset_name, savefile_name, x = x_value, y = y_value)
}
