% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.plot.R
\name{create_plot}
\alias{create_plot}
\title{This function creates a scatter plot using designated x and y values and saves the plot to a file.}
\usage{
create_plot(filepath, savefile)
}
\arguments{
\item{filepath}{the name of a dataset that has been loaded into the
environment or its corresponding file path}

\item{savefile}{the name that your plot file will be saved as}

\item{x}{= x value}

\item{y}{= y value}
}
\value{
If non-integer values are entered for the plot function, console will print "Error: Values must be integers!"
}
\description{
This function creates a scatter plot using designated x and y values and saves the plot to a file.
}
