% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cleaner.R
\name{data_cleaning}
\alias{data_cleaning}
\title{This function cleans all of the NA values from a data set.}
\usage{
data_cleaning(filepath)
}
\arguments{
\item{filepath}{the name of a dataset that has been loaded into the
environment or its corresponding file path}
}
\value{
Nothing will print if data cleaning is successful.
If unsuccessful, console will print "NAs still present!").
}
\description{
This function cleans all of the NA values from a data set.
}
\examples{
data_cleaning(dataset_name)
}
